Extract text from extract_assets.py
