+++ Makefile
@@ -318,8 +318,6 @@ setup:
 
 assets:
 	python3 extract_assets.py -j $(N_THREADS) -Z Wno-hardcoded-pointer
-	python3 tools/msg/dis/msgdisNES.py -o assets/text/message_data.h
-	python3 tools/msg/dis/msgdisStaff.py -o assets/text/staff_message_data.h
 
 ## Assembly generation
 disasm:
@@ -369,10 +367,10 @@ build/data/%.o: data/%.s
 	$(AS) $(ASFLAGS) $< -o $@
 
 build/assets/text/message_data.enc.h: assets/text/message_data.h
-	python3 tools/msg/enc/msgencNES.py -o $@ $<
+	python3 tools/msg/nes/msgencNES.py -o $@ $<
 
 build/assets/text/staff_message_data.enc.h: assets/text/staff_message_data.h
-	python3 tools/msg/enc/msgencStaff.py -o $@ $<
+	python3 tools/msg/staff/msgencStaff.py -o $@ $<
 
 build/assets/text/message_data_static.o: build/assets/text/message_data.enc.h
 build/assets/text/staff_message_data_static.o: build/assets/text/staff_message_data.enc.h
